requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
openai

python-multipart

pytesseract
pillow
pdfplumber
python-docx
beautifulsoup4
trafilatura

